Corrections linter
