+++ src/components/custom/widget/EmploiDuTempsWidget.tsx
@@ -1,16 +1,17 @@
-import { TextSkeleton } from "@/components/Skeleton";
 import { useTheme } from "@/contexts/ThemeContext";
 import { useEmploiDuTemps } from "@/hooks/useEmploiDuTemps";
 import type { AppStackParamList } from "@/services/storage/types";
 import { useNavigation } from "@react-navigation/native";
 import type { StackNavigationProp } from "@react-navigation/stack";
 
+import { EmploiDuTempsWidgetCourse } from "@/components/custom/widget/EmploiDuTempsWidgetCourse";
+import { EmploiDuTempsWidgetLoading } from "@/components/custom/widget/EmploiDuTempsWidgetLoading";
 import { useAuth } from "@/hooks/account/useAuth";
 import type { Course } from "@/types/emploiDuTemps";
 import { useTranslation } from "react-i18next";
 import { Text, TouchableOpacity, View } from "react-native";
 
-type AppScreenNavigationProp = StackNavigationProp<AppStackParamList>;
+export type AppScreenNavigationProp = StackNavigationProp<AppStackParamList>;
 
 export function EmploiDuTempsWidget() {
   const { t } = useTranslation();
@@ -173,40 +174,7 @@ export function EmploiDuTempsWidget() {
         className="rounded-lg flex flex-col gap-3"
       >
         {filteredCourses?.map((course: Course) => (
-          <View
-            key={course.id}
-            className="flex flex-col rounded-lg gap-1.5 py-2"
-            style={{ backgroundColor: theme.card }}
-          >
-            <Text
-              className="text-base ml-4"
-              style={{ color: theme.text }}
-              ellipsizeMode="tail"
-            >
-              {course.titre}
-            </Text>
-            <View className="flex flex-row items-center gap-2">
-              <Text
-                className="text-sm ml-4"
-                style={{ color: theme.text }}
-                ellipsizeMode="tail"
-              >
-                {course.heure_debut} - {course.heure_fin}
-              </Text>
-              <View>
-                <Text
-                  className="pl-1 pr-1 rounded-md text-base ml-4"
-                  style={{
-                    backgroundColor: theme.primary,
-                    color: theme.background,
-                  }}
-                  ellipsizeMode="tail"
-                >
-                  {course.salles}
-                </Text>
-              </View>
-            </View>
-          </View>
+          <EmploiDuTempsWidgetCourse key={course.id} course={course} />
         ))}
       </TouchableOpacity>
     </View>
@@ -214,45 +182,3 @@ export function EmploiDuTempsWidget() {
 }
 
 export default EmploiDuTempsWidget;
-
-export const EmploiDuTempsWidgetLoading = () => {
-  const { theme } = useTheme();
-  const navigation = useNavigation<AppScreenNavigationProp>();
-
-  const skeletonCount = () => Math.floor(Math.random() * 3) + 1;
-
-  return (
-    <View className="flex flex-col gap-2">
-      <TextSkeleton lines={1} lastLineWidth={128} />
-      <TouchableOpacity
-        onPress={() => navigation.navigate("EmploiDuTemps")}
-        className="px-6 py-4 rounded-lg flex flex-col gap-6"
-        style={{ backgroundColor: theme.card }}
-      >
-        <View className="flex flex-col gap-2">
-          {[...Array(skeletonCount()).keys()].map((index) => (
-            <TextSkeleton lines={1} key={index} />
-          ))}
-        </View>
-
-        <View className="flex flex-col gap-2">
-          {[...Array(skeletonCount()).keys()].map((index) => (
-            <TextSkeleton lines={1} key={index} />
-          ))}
-        </View>
-
-        <View className="flex flex-col gap-2">
-          {[...Array(skeletonCount()).keys()].map((index) => (
-            <TextSkeleton lines={1} key={index} />
-          ))}
-        </View>
-
-        <View className="flex flex-col gap-2">
-          {[...Array(skeletonCount()).keys()].map((index) => (
-            <TextSkeleton lines={1} key={index} />
-          ))}
-        </View>
-      </TouchableOpacity>
-    </View>
-  );
-};
