+++ src/hooks/useEmploiDuTemps.ts
@@ -1,11 +1,12 @@
 import { getEmploiDuTempsToday } from "@/lib/emploiDuTemps";
 import { QUERY_KEYS } from "@/lib/queryKeys";
-import { useQuery } from "@tanstack/react-query";
+import { skipToken, useQuery } from "@tanstack/react-query";
 
-export const useEmploiDuTemps = (email: string) => {
+export const useEmploiDuTemps = (email?: string) => {
   const { data, isPending, refetch, error, isError } = useQuery({
     queryKey: QUERY_KEYS.emploiDuTemps,
-    queryFn: () => getEmploiDuTempsToday(email),
+    queryFn: email ? () => getEmploiDuTempsToday(email) : skipToken,
+    enabled: !!email,
   });
 
   return {
